This is put up to give people an idea (at best) of the direction I'm going with the medialibrary. Nothing is remotely close to ready.
The interesting parts are the new DynamicMedia plug in around the Data and View folders

All data is now retrieved via a data provider in the data folder (which need to be expanded upon later for filtering / sorting).

In the view a IDataSourceList type object is returned now instead of a List which inherits IList<T> that has to be bound to the itemsource.

Also the was edited to enable virtualization of the data.