OK this is more just a WIP build and has changes which will be dropped, reworked and all sorts of horrible things.
On the plus side it shows the Hello World plugin loading quickly with 10000 items but as you scroll it crawls (still heaps faster than before). I suspect I'll have to extend the VirtualizingPanel to take advantage of the isLoading property on each item but I haven't started down that path yet.

An extra DLL is required which isn't apart of this commit so not worth testing or building against as it'll change.